Restructuring core library
