+++ lib/compiler/lexer.js
@@ -1467,6 +1467,8 @@
 					offset += this._indentStyle.length;
 				} else if (offset == whitespace.length) {
 					break;
+				} else if (this._opts.silent) {
+					break;
 				} else {
 					// workaround to report correct location
 					this._loc += indent.length - whitespace.length;
