--- lib/imba/core.events.js
@@ -1,82 +1,0 @@
-(function(){
-	
-	
-	function emit__(event,args,node){
-		// var node = cbs[event]
-		var prev,cb,ret;
-		
-		while ((prev = node) && (node = node.next)){
-			if (cb = node.listener) {
-				if (node.path && cb[node.path]) {
-					ret = args ? (cb[node.path].apply(cb,args)) : (cb[node.path]());
-				} else {
-					// check if it is a method?
-					ret = args ? (cb.apply(node,args)) : (cb.call(node));
-				};
-			};
-			
-			if (node.times && --node.times <= 0) {
-				prev.next = node.next;
-				node.listener = null;
-			};
-		};
-		return;
-	};
-	
-	// method for registering a listener on object
-	Imba.listen = function (obj,event,listener,path){
-		var $1;
-		var cbs,list,tail;
-		cbs = obj.__listeners__ || (obj.__listeners__ = {});
-		list = cbs[($1 = event)] || (cbs[$1] = {});
-		tail = list.tail || (list.tail = (list.next = {}));
-		tail.listener = listener;
-		tail.path = path;
-		list.tail = tail.next = {};
-		return tail;
-	};
-	
-	Imba.once = function (obj,event,listener){
-		var tail = Imba.listen(obj,event,listener);
-		tail.times = 1;
-		return tail;
-	};
-	
-	Imba.unlisten = function (obj,event,cb,meth){
-		var node,prev;
-		var meta = obj.__listeners__;
-		if (!meta) { return };
-		
-		if (node = meta[event]) {
-			while ((prev = node) && (node = node.next)){
-				if (node == cb || node.listener == cb) {
-					prev.next = node.next;
-					// check for correct path as well?
-					node.listener = null;
-					break;
-				};
-			};
-		};
-		return;
-	};
-	
-	Imba.emit = function (obj,event,params){
-		var cb;
-		if (cb = obj.__listeners__) {
-			if (cb[event]) { emit__(event,params,cb[event]) };
-			if (cb.all) { emit__(event,[event,params],cb.all) }; // and event != 'all'
-		};
-		return;
-	};
-	
-	return Imba.observeProperty = function (observer,key,trigger,target,prev){
-		if (prev && typeof prev == 'object') {
-			Imba.unlisten(prev,'all',observer,trigger);
-		};
-		if (target && typeof target == 'object') {
-			Imba.listen(target,'all',observer,trigger);
-		};
-		return this;
-	};
-
-})();
