+++ lib/imba/dom/event-manager.js
@@ -1,0 +1,150 @@
+(function(){
+	function iter$(a){ return a ? (a.toArray ? a.toArray() : a) : []; };
+	/*
+	
+	Manager for listening to and delegating events in Imba. A single instance
+	is always created by Imba (as `Imba.Events`), which handles and delegates all
+	events at the very root of the document. Imba does not capture all events
+	by default, so if you want to make sure exotic or custom DOMEvents are delegated
+	in Imba you will need to register them in `Imba.Events.register(myCustomEventName)`
+	
+	@iname manager
+	
+	*/
+	
+	Imba.EventManager = function EventManager(node,pars){
+		var self = this;
+		if(!pars||pars.constructor !== Object) pars = {};
+		var events = pars.events !== undefined ? pars.events : [];
+		self.setRoot(node);
+		self.setCount(0);
+		self.setListeners([]);
+		self.setDelegators({});
+		self.setDelegator(function(e) {
+			// console.log "delegating event?! {e}"
+			self.delegate(e);
+			return true;
+		});
+		
+		for (var i = 0, ary = iter$(events), len = ary.length; i < len; i++) {
+			self.register(ary[i]);
+		};
+		
+		return self;
+	};
+	
+	Imba.EventManager.prototype.root = function(v){ return this._root; }
+	Imba.EventManager.prototype.setRoot = function(v){ this._root = v; return this; };
+	Imba.EventManager.prototype.count = function(v){ return this._count; }
+	Imba.EventManager.prototype.setCount = function(v){ this._count = v; return this; };
+	Imba.EventManager.prototype.__enabled = {'default': false,watch: 'enabledDidSet',name: 'enabled'};
+	Imba.EventManager.prototype.enabled = function(v){ return this._enabled; }
+	Imba.EventManager.prototype.setEnabled = function(v){
+		var a = this.enabled();
+		if(v != a) { this._enabled = v; }
+		if(v != a) { this.enabledDidSet && this.enabledDidSet(v,a,this.__enabled) }
+		return this;
+	}
+	Imba.EventManager.prototype._enabled = false;
+	Imba.EventManager.prototype.listeners = function(v){ return this._listeners; }
+	Imba.EventManager.prototype.setListeners = function(v){ this._listeners = v; return this; };
+	Imba.EventManager.prototype.delegators = function(v){ return this._delegators; }
+	Imba.EventManager.prototype.setDelegators = function(v){ this._delegators = v; return this; };
+	Imba.EventManager.prototype.delegator = function(v){ return this._delegator; }
+	Imba.EventManager.prototype.setDelegator = function(v){ this._delegator = v; return this; };
+	
+	Imba.EventManager.prototype.enabledDidSet = function (bool){
+		bool ? (this.onenable()) : (this.ondisable());
+		return this;
+	};
+	
+	/*
+	
+		Tell the current EventManager to intercept and handle event of a certain name.
+		By default, Imba.Events will register interceptors for: *keydown*, *keyup*, 
+		*keypress*, *textInput*, *input*, *change*, *submit*, *focusin*, *focusout*, 
+		*blur*, *contextmenu*, *dblclick*, *mousewheel*, *wheel*
+	
+		*/
+	
+	Imba.EventManager.prototype.register = function (name,handler){
+		if(handler === undefined) handler = true;
+		if (name instanceof Array) {
+			for (var i = 0, ary = iter$(name), len = ary.length; i < len; i++) {
+				this.register(ary[i],handler);
+			};
+			return this;
+		};
+		
+		if (this.delegators()[name]) { return this };
+		// console.log("register for event {name}")
+		var fn = this.delegators()[name] = handler instanceof Function ? (handler) : (this.delegator());
+		if (this.enabled()) { return this.root().addEventListener(name,fn,true) };
+	};
+	
+	Imba.EventManager.prototype.listen = function (name,handler,capture){
+		if(capture === undefined) capture = true;
+		this.listeners().push([name,handler,capture]);
+		if (this.enabled()) { this.root().addEventListener(name,handler,capture) };
+		return this;
+	};
+	
+	Imba.EventManager.prototype.delegate = function (e){
+		this.setCount(this.count() + 1);
+		var event = Imba.Event.wrap(e);
+		event.process();
+		return this;
+	};
+	
+	/*
+	
+		Create a new Imba.Event
+	
+		*/
+	
+	Imba.EventManager.prototype.create = function (type,target,pars){
+		if(!pars||pars.constructor !== Object) pars = {};
+		var data = pars.data !== undefined ? pars.data : null;
+		var source = pars.source !== undefined ? pars.source : null;
+		var event = Imba.Event.wrap({type: type,target: target});
+		if (data) { (event.setData(data),data) };
+		if (source) { (event.setSource(source),source) };
+		return event;
+	};
+	
+	/*
+	
+		Trigger / process an Imba.Event.
+	
+		*/
+	
+	Imba.EventManager.prototype.trigger = function (){
+		return this.create.apply(this,arguments).process();
+	};
+	
+	Imba.EventManager.prototype.onenable = function (){
+		for (var o = this.delegators(), i = 0, keys = Object.keys(o), l = keys.length; i < l; i++){
+			this.root().addEventListener(keys[i],o[keys[i]],true);
+		};
+		
+		for (var i = 0, ary = iter$(this.listeners()), len = ary.length, item; i < len; i++) {
+			item = ary[i];
+			this.root().addEventListener(item[0],item[1],item[2]);
+		};
+		return this;
+	};
+	
+	Imba.EventManager.prototype.ondisable = function (){
+		for (var o = this.delegators(), i = 0, keys = Object.keys(o), l = keys.length; i < l; i++){
+			this.root().removeEventListener(keys[i],o[keys[i]],true);
+		};
+		
+		for (var i = 0, ary = iter$(this.listeners()), len = ary.length, item; i < len; i++) {
+			item = ary[i];
+			this.root().removeEventListener(item[0],item[1],item[2]);
+		};
+		return this;
+	};
+	return Imba.EventManager;
+
+})();
