+++ lib/imba/dom/server.js
@@ -1,5 +1,4 @@
 (function(){
-	var global_;
 	
 	
 	// TODO classes should not be global,
@@ -217,17 +216,9 @@
 	
 	tag$.extendTag('element', function(tag){
 		
-		tag.prototype.toString = function (){
-			return this.dom().toString();
-		};
-	});
-	
-	tag$.extendTag('htmlelement', function(tag){
-		
 		tag.prototype.empty = function (){
 			this._dom.children = [];
 			this._dom.innerHTML = null;
-			// @dom.removeChild(@dom:firstChild) while @dom:firstChild
 			this._empty = true;
 			return this;
 		};
@@ -243,10 +234,13 @@
 		tag.prototype.prepend = function (item){
 			return this._dom.children.unshift(item);
 		};
+		
+		tag.prototype.toString = function (){
+			return this.dom().toString();
+		};
 	});
 	
-	
-	tag$.extendTag('html', function(tag){
+	return tag$.extendTag('html', function(tag){
 		
 		tag.prototype.doctype = function (){
 			return this._doctype || "<!doctype html>";
@@ -254,17 +248,10 @@
 		
 		tag.prototype.toString = function (){
 			return this.doctype() + tag.__super__.toString.apply(this,arguments);
-			// <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
 		};
 	});
 	
-	tag$.extendTag('style', function(tag){
-		
-		tag.prototype.toString = function (){
-			return "<style/>";
-		};
-	});
 	
-	return global.document || (global.document = Imba.document());
+	// global:document ||= Imba.document
 
 })();
