+++ lib/imba/dom/svg.js
