+++ lib/imba/dom/tag.js
@@ -7,11 +7,6 @@
 	
 	Imba.CSSKeyMap = {};
 	
-	Imba.static = function (items,nr){
-		items.static = nr;
-		return items;
-	};
-	
 	/*
 	Get the current document
 	*/
@@ -31,6 +26,14 @@
 	Imba.root = function (){
 		return tag$wrap(Imba.document().body);
 	};
+	
+	
+	Imba.static = function (items,nr){
+		items.static = nr;
+		return items;
+	};
+	
+	
 	
 	/*
 	This is the baseclass that all tags in imba inherit from.
