+++ lib/imba/dom/touch.js
@@ -1,0 +1,538 @@
+(function(){
+	function iter$(a){ return a ? (a.toArray ? a.toArray() : a) : []; };
+	// Imba.Touch
+	// Began	A finger touched the screen.
+	// Moved	A finger moved on the screen.
+	// Stationary	A finger is touching the screen but hasn't moved.
+	// Ended	A finger was lifted from the screen. This is the final phase of a touch.
+	// Canceled The system cancelled tracking for the touch.
+	
+	/*
+	Consolidates mouse and touch events. Touch objects persist across a touch,
+	from touchstart until end/cancel. When a touch starts, it will traverse
+	down from the innermost target, until it finds a node that responds to
+	ontouchstart. Unless the touch is explicitly redirected, the touch will
+	call ontouchmove and ontouchend / ontouchcancel on the responder when appropriate.
+	
+		tag draggable
+			# called when a touch starts
+			def ontouchstart touch
+				flag 'dragging'
+				self
+			
+			# called when touch moves - same touch object
+			def ontouchmove touch
+				# move the node with touch
+				css top: touch.dy, left: touch.dx
+			
+			# called when touch ends
+			def ontouchend touch
+				unflag 'dragging'
+	
+	@iname touch
+	*/
+	
+	Imba.Touch = function Touch(event,pointer){
+		// @native  = false
+		this.setEvent(event);
+		this.setData({});
+		this.setActive(true);
+		this._button = event && event.button || 0;
+		this._suppress = false; // deprecated
+		this._captured = false;
+		this.setBubble(false);
+		pointer = pointer;
+		this.setUpdates(0);
+		return this;
+	};
+	
+	Imba.Touch.LastTimestamp = 0;
+	Imba.Touch.TapTimeout = 50;
+	
+	// var lastNativeTouchTimeout = 50
+	
+	var touches = [];
+	var count = 0;
+	var identifiers = {};
+	
+	Imba.Touch.count = function (){
+		return count;
+	};
+	
+	Imba.Touch.lookup = function (item){
+		return item && (item.__touch__ || identifiers[item.identifier]);
+	};
+	
+	Imba.Touch.release = function (item,touch){
+		var v_, $1;
+		(((v_ = identifiers[item.identifier]),delete identifiers[item.identifier], v_));
+		((($1 = item.__touch__),delete item.__touch__, $1));
+		return;
+	};
+	
+	Imba.Touch.ontouchstart = function (e){
+		for (var i = 0, ary = iter$(e.changedTouches), len = ary.length, t; i < len; i++) {
+			t = ary[i];
+			if (this.lookup(t)) { continue; };
+			var touch = identifiers[t.identifier] = new this(e); // (e)
+			t.__touch__ = touch;
+			touches.push(touch);
+			count++;
+			touch.touchstart(e,t);
+		};
+		return this;
+	};
+	
+	Imba.Touch.ontouchmove = function (e){
+		var touch;
+		for (var i = 0, ary = iter$(e.changedTouches), len = ary.length, t; i < len; i++) {
+			t = ary[i];
+			if (touch = this.lookup(t)) {
+				touch.touchmove(e,t);
+			};
+		};
+		
+		return this;
+	};
+	
+	Imba.Touch.ontouchend = function (e){
+		var touch;
+		for (var i = 0, ary = iter$(e.changedTouches), len = ary.length, t; i < len; i++) {
+			t = ary[i];
+			if (touch = this.lookup(t)) {
+				touch.touchend(e,t);
+				this.release(t,touch);
+				count--;
+			};
+		};
+		
+		// e.preventDefault
+		// not always supported!
+		// touches = touches.filter(||)
+		return this;
+	};
+	
+	Imba.Touch.ontouchcancel = function (e){
+		var touch;
+		for (var i = 0, ary = iter$(e.changedTouches), len = ary.length, t; i < len; i++) {
+			t = ary[i];
+			if (touch = this.lookup(t)) {
+				touch.touchcancel(e,t);
+				this.release(t,touch);
+				count--;
+			};
+		};
+		return this;
+	};
+	
+	Imba.Touch.onmousedown = function (e){
+		return this;
+	};
+	
+	Imba.Touch.onmousemove = function (e){
+		return this;
+	};
+	
+	Imba.Touch.onmouseup = function (e){
+		return this;
+	};
+	
+	
+	Imba.Touch.prototype.phase = function(v){ return this._phase; }
+	Imba.Touch.prototype.setPhase = function(v){ this._phase = v; return this; };
+	Imba.Touch.prototype.active = function(v){ return this._active; }
+	Imba.Touch.prototype.setActive = function(v){ this._active = v; return this; };
+	Imba.Touch.prototype.event = function(v){ return this._event; }
+	Imba.Touch.prototype.setEvent = function(v){ this._event = v; return this; };
+	Imba.Touch.prototype.pointer = function(v){ return this._pointer; }
+	Imba.Touch.prototype.setPointer = function(v){ this._pointer = v; return this; };
+	Imba.Touch.prototype.target = function(v){ return this._target; }
+	Imba.Touch.prototype.setTarget = function(v){ this._target = v; return this; };
+	Imba.Touch.prototype.handler = function(v){ return this._handler; }
+	Imba.Touch.prototype.setHandler = function(v){ this._handler = v; return this; };
+	Imba.Touch.prototype.updates = function(v){ return this._updates; }
+	Imba.Touch.prototype.setUpdates = function(v){ this._updates = v; return this; };
+	Imba.Touch.prototype.suppress = function(v){ return this._suppress; }
+	Imba.Touch.prototype.setSuppress = function(v){ this._suppress = v; return this; };
+	Imba.Touch.prototype.data = function(v){ return this._data; }
+	Imba.Touch.prototype.setData = function(v){ this._data = v; return this; };
+	Imba.Touch.prototype.__bubble = {chainable: true,name: 'bubble'};
+	Imba.Touch.prototype.bubble = function(v){ return v !== undefined ? (this.setBubble(v),this) : this._bubble; }
+	Imba.Touch.prototype.setBubble = function(v){ this._bubble = v; return this; };
+	
+	Imba.Touch.prototype.gestures = function(v){ return this._gestures; }
+	Imba.Touch.prototype.setGestures = function(v){ this._gestures = v; return this; };
+	
+	/*
+		@internal
+		@constructor
+		*/
+	
+	Imba.Touch.prototype.capture = function (){
+		this._captured = true;
+		this._event && this._event.preventDefault();
+		return this;
+	};
+	
+	Imba.Touch.prototype.isCaptured = function (){
+		return !!this._captured;
+	};
+	
+	/*
+		Extend the touch with a plugin / gesture. 
+		All events (touchstart,move etc) for the touch
+		will be triggered on the plugins in the order they
+		are added.
+		*/
+	
+	Imba.Touch.prototype.extend = function (plugin){
+		// console.log "added gesture!!!"
+		this._gestures || (this._gestures = []);
+		this._gestures.push(plugin);
+		return this;
+	};
+	
+	/*
+		Redirect touch to specified target. ontouchstart will always be
+		called on the new target.
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.redirect = function (target){
+		this._redirect = target;
+		return this;
+	};
+	
+	/*
+		Suppress the default behaviour. Will call preventDefault for
+		all native events that are part of the touch.
+		*/
+	
+	Imba.Touch.prototype.suppress = function (){
+		// collision with the suppress property
+		this._active = false;
+		return this;
+	};
+	
+	Imba.Touch.prototype.setSuppress = function (value){
+		console.warn('Imba.Touch#suppress= is deprecated');
+		this._supress = value;
+		this;
+		return this;
+	};
+	
+	Imba.Touch.prototype.touchstart = function (e,t){
+		this._event = e;
+		this._touch = t;
+		this._button = 0;
+		this._x = t.clientX;
+		this._y = t.clientY;
+		this.began();
+		if (e && this.isCaptured()) { e.preventDefault() };
+		return this;
+	};
+	
+	Imba.Touch.prototype.touchmove = function (e,t){
+		this._event = e;
+		this._x = t.clientX;
+		this._y = t.clientY;
+		this.update();
+		if (e && this.isCaptured()) { e.preventDefault() };
+		return this;
+	};
+	
+	Imba.Touch.prototype.touchend = function (e,t){
+		this._event = e;
+		this._x = t.clientX;
+		this._y = t.clientY;
+		this.ended();
+		
+		Imba.Touch.LastTimestamp = e.timeStamp;
+		
+		if (this._maxdr < 20) {
+			var tap = new Imba.Event(e);
+			tap.setType('tap');
+			tap.process();
+			if (tap._responder) { e.preventDefault() };
+		};
+		
+		if (e && this.isCaptured()) {
+			e.preventDefault();
+		};
+		
+		return this;
+	};
+	
+	Imba.Touch.prototype.touchcancel = function (e,t){
+		return this.cancel();
+	};
+	
+	Imba.Touch.prototype.mousedown = function (e,t){
+		var self = this;
+		self._event = e;
+		self._button = e.button;
+		self._x = t.clientX;
+		self._y = t.clientY;
+		self.began();
+		
+		self._mousemove = function(e) { return self.mousemove(e,e); };
+		Imba.document().addEventListener('mousemove',self._mousemove,true);
+		return self;
+	};
+	
+	Imba.Touch.prototype.mousemove = function (e,t){
+		this._x = t.clientX;
+		this._y = t.clientY;
+		this._event = e;
+		if (this.isCaptured()) { e.preventDefault() };
+		this.update();
+		this.move();
+		return this;
+	};
+	
+	Imba.Touch.prototype.mouseup = function (e,t){
+		this._x = t.clientX;
+		this._y = t.clientY;
+		this.ended();
+		Imba.document().removeEventListener('mousemove',this._mousemove,true);
+		this._mousemove = null;
+		return this;
+	};
+	
+	Imba.Touch.prototype.idle = function (){
+		return this.update();
+	};
+	
+	Imba.Touch.prototype.began = function (){
+		this._maxdr = this._dr = 0;
+		this._x0 = this._x;
+		this._y0 = this._y;
+		
+		var dom = this.event().target;
+		var node = null;
+		
+		this._sourceTarget = dom && tag$wrap(dom);
+		
+		while (dom){
+			node = tag$wrap(dom);
+			if (node && node.ontouchstart) {
+				this._bubble = false;
+				this.setTarget(node);
+				this.target().ontouchstart(this);
+				if (!this._bubble) { break; };
+			};
+			dom = dom.parentNode;
+		};
+		
+		this._updates++;
+		return this;
+	};
+	
+	Imba.Touch.prototype.update = function (){
+		var target_;
+		if (!this._active) { return this };
+		
+		var dr = Math.sqrt(this.dx() * this.dx() + this.dy() * this.dy());
+		if (dr > this._dr) { this._maxdr = dr };
+		this._dr = dr;
+		
+		// catching a touch-redirect?!?
+		if (this._redirect) {
+			if (this._target && this._target.ontouchcancel) {
+				this._target.ontouchcancel(this);
+			};
+			this.setTarget(this._redirect);
+			this._redirect = null;
+			if (this.target().ontouchstart) { this.target().ontouchstart(this) };
+		};
+		
+		
+		this._updates++;
+		if (this._gestures) {
+			for (var i = 0, ary = iter$(this._gestures), len = ary.length; i < len; i++) {
+				ary[i].ontouchupdate(this);
+			};
+		};
+		
+		(target_ = this.target()) && target_.ontouchupdate  &&  target_.ontouchupdate(this);
+		return this;
+	};
+	
+	Imba.Touch.prototype.move = function (){
+		var target_;
+		if (!this._active) { return this };
+		
+		if (this._gestures) {
+			for (var i = 0, ary = iter$(this._gestures), len = ary.length, g; i < len; i++) {
+				g = ary[i];
+				if (g.ontouchmove) { g.ontouchmove(this,this._event) };
+			};
+		};
+		
+		(target_ = this.target()) && target_.ontouchmove  &&  target_.ontouchmove(this,this._event);
+		return this;
+	};
+	
+	Imba.Touch.prototype.ended = function (){
+		var target_;
+		if (!this._active) { return this };
+		
+		this._updates++;
+		
+		if (this._gestures) {
+			for (var i = 0, ary = iter$(this._gestures), len = ary.length; i < len; i++) {
+				ary[i].ontouchend(this);
+			};
+		};
+		
+		(target_ = this.target()) && target_.ontouchend  &&  target_.ontouchend(this);
+		
+		return this;
+	};
+	
+	Imba.Touch.prototype.cancel = function (){
+		if (!this._cancelled) {
+			this._cancelled = true;
+			this.cancelled();
+			if (this._mousemove) { Imba.document().removeEventListener('mousemove',this._mousemove,true) };
+		};
+		return this;
+	};
+	
+	Imba.Touch.prototype.cancelled = function (){
+		var target_;
+		if (!this._active) { return this };
+		
+		this._cancelled = true;
+		this._updates++;
+		
+		if (this._gestures) {
+			for (var i = 0, ary = iter$(this._gestures), len = ary.length, g; i < len; i++) {
+				g = ary[i];
+				if (g.ontouchcancel) { g.ontouchcancel(this) };
+			};
+		};
+		
+		(target_ = this.target()) && target_.ontouchcancel  &&  target_.ontouchcancel(this);
+		return this;
+	};
+	
+	/*
+		The absolute distance the touch has moved from starting position 
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.dr = function (){
+		return this._dr;
+	};
+	
+	/*
+		The distance the touch has moved horizontally
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.dx = function (){
+		return this._x - this._x0;
+	};
+	
+	/*
+		The distance the touch has moved vertically
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.dy = function (){
+		return this._y - this._y0;
+	};
+	
+	/*
+		Initial horizontal position of touch
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.x0 = function (){
+		return this._x0;
+	};
+	
+	/*
+		Initial vertical position of touch
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.y0 = function (){
+		return this._y0;
+	};
+	
+	/*
+		Horizontal position of touch
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.x = function (){
+		return this._x;
+	};
+	
+	/*
+		Vertical position of touch
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.y = function (){
+		return this._y;
+	};
+	
+	/*
+		Horizontal position of touch relative to target
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.tx = function (){
+		this._targetBox || (this._targetBox = this._target.dom().getBoundingClientRect());
+		return this._x - this._targetBox.left;
+	};
+	
+	/*
+		Vertical position of touch relative to target
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.ty = function (){
+		this._targetBox || (this._targetBox = this._target.dom().getBoundingClientRect());
+		return this._y - this._targetBox.top;
+	};
+	
+	/*
+		Button pressed in this touch. Native touches defaults to left-click (0)
+		@return {Number}
+		*/
+	
+	Imba.Touch.prototype.button = function (){
+		return this._button;
+	}; // @pointer ? @pointer.button : 0
+	
+	Imba.Touch.prototype.sourceTarget = function (){
+		return this._sourceTarget;
+	};
+	
+	
+	Imba.TouchGesture = function TouchGesture(){ };
+	
+	Imba.TouchGesture.prototype.__active = {'default': false,name: 'active'};
+	Imba.TouchGesture.prototype.active = function(v){ return this._active; }
+	Imba.TouchGesture.prototype.setActive = function(v){ this._active = v; return this; }
+	Imba.TouchGesture.prototype._active = false;
+	
+	Imba.TouchGesture.prototype.ontouchstart = function (e){
+		return this;
+	};
+	
+	Imba.TouchGesture.prototype.ontouchupdate = function (e){
+		return this;
+	};
+	
+	Imba.TouchGesture.prototype.ontouchend = function (e){
+		return this;
+	};
+	return Imba.TouchGesture;
+	
+
+})();
