+++ src/imba/dom/event-manager.imba
@@ -1,0 +1,102 @@
+###
+
+Manager for listening to and delegating events in Imba. A single instance
+is always created by Imba (as `Imba.Events`), which handles and delegates all
+events at the very root of the document. Imba does not capture all events
+by default, so if you want to make sure exotic or custom DOMEvents are delegated
+in Imba you will need to register them in `Imba.Events.register(myCustomEventName)`
+
+@iname manager
+
+###
+class Imba.EventManager
+
+	prop root
+	prop count
+	prop enabled default: no, watch: yes
+	prop listeners
+	prop delegators
+	prop delegator
+
+	def enabled-did-set bool
+		bool ? onenable : ondisable
+		self
+
+	def initialize node, events: []
+		root = node
+		count = 0
+		listeners = []
+		delegators = {}
+		delegator = do |e| 
+			# console.log "delegating event?! {e}"
+			delegate(e)
+			return true
+
+		for event in events
+			register(event)
+
+		return self
+
+	###
+
+	Tell the current EventManager to intercept and handle event of a certain name.
+	By default, Imba.Events will register interceptors for: *keydown*, *keyup*, 
+	*keypress*, *textInput*, *input*, *change*, *submit*, *focusin*, *focusout*, 
+	*blur*, *contextmenu*, *dblclick*, *mousewheel*, *wheel*
+
+	###
+	def register name, handler = true
+		if name isa Array
+			register(v,handler) for v in name
+			return self
+
+		return self if delegators[name]
+		# console.log("register for event {name}")
+		var fn = delegators[name] = handler isa Function ? handler : delegator
+		root.addEventListener(name,fn,yes) if enabled
+
+	def listen name, handler, capture = yes
+		listeners.push([name,handler,capture])
+		root.addEventListener(name,handler,capture) if enabled
+		self
+
+	def delegate e
+		count += 1
+		var event = Imba.Event.wrap(e)
+		event.process
+		self
+
+	###
+
+	Create a new Imba.Event
+
+	###
+	def create type, target, data: null, source: null
+		var event = Imba.Event.wrap type: type, target: target
+		event.data = data if data
+		event.source = source if source
+		event
+
+	###
+
+	Trigger / process an Imba.Event.
+
+	###
+	def trigger
+		create(*arguments).process
+
+	def onenable
+		for own name,handler of delegators
+			root.addEventListener(name,handler,yes)
+
+		for item in listeners
+			root.addEventListener(item[0],item[1],item[2])
+		self
+
+	def ondisable
+		for own name,handler of delegators
+			root.removeEventListener(name,handler,yes)
+
+		for item in listeners
+			root.removeEventListener(item[0],item[1],item[2])
+		self
