+++ src/imba/dom/svg.imba
